Rename files to gum-gum IX format
