+++ package.json
@@ -2,7 +2,7 @@
   "name": "gumgum-module",
   "version": "2.1.0",
   "description": "a complete self-serve partner development package for the header tag wrapper",
-  "main": "gumgum-htb.js",
+  "main": "gum-gum-htb.js",
   "scripts": {
     "test": "jasmine"
   },
